Delta Temperature

The 1979-2000 rearing temperature in the North and South Delta in °C

delta_temperature

Format

a 3 dimensional array [12 months, 22 years, 2 deltas]

Source

Data Wrangling and Additional Temperature Modeling: Sadie Gill sgill@flowwest.com

Details

Water temperature in the Deltas were modeled as a function of air temperature.

North Delta was modeled using a sample of recent measured water temperature at CDEC station EMM and NOAA CDO air temperature data from Antioch.

The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

South Delta was modeled using a sample of recent measured water temperature at CDEC station MTB and NOAA CDO air temperature data Stockton Airport.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River below Vernalis.

Air Temperature Data Source:

NOAA Climate Data Online (CDO) accessed using the rnoaa R package developed by rOpenSci.

Stream Water Temperature Data Sources:

DWR California Data Exchange Center (CDEC) accessed using the CDECRetrieve R packaged developed by FlowWest.