Migratory Corridor Water Temperature

The median proportion of days over 20°C per month

migratory_temperature_proportion_over_20

Format

a 31 by 12 matrix [watershed by month]

Source

Method developed by James T. Peterson jt.peterson@oregonstate.edu

Data wrangling by Sadie Gill sgill@flowwest.com

Details

Watersheds were assigned membership to one of the following three locations to represent the migratory corridor temperature conditions. The data was summarised by calculating the median proportion of days over 20°C for each month during the available period of record.

Sacramento River, Sacramento River Below Wilkins Slough: Upper Sacramento River, Antelope Creek, Battle Creek, Bear Creek, Big Chico Creek Butte Creek, Clear Creek, Cottonwood Creek, Cow Creek, Deer Creek, Elder Creek, Mill Creek, Paynes Creek, Stony Creek, Thomes Creek, Bear River, Feather River, Yuba River American River

South Delta, Mokelumne River At San Joaquin River: Calaveras River, Cosumnes River, Mokelumne River

San Joaquin River, San Joaquin River Near Vernalis: Merced River, Stanislaus River, Tuolumne River

No spawning in these regions (assigned value of zero): Upper-mid Sacramento River, Sutter Bypass, Lower-mid Sacramento River, Yolo Bypass, Lower Sacramento River, San Joaquin River